move assertable to maplibre helper
