+++ projects/showcase/cypress/support/e2e-driver.ts
@@ -1,55 +1,6 @@
-/// <reference types="cypress" />
-import { Assertable, then } from "@shellygo/cypress-test-utils/assertable";
-import pixelmatch from "pixelmatch";
 import { PNG } from "pngjs";
-import { MaplibreCypressHelper } from "./maplibre-helper";
+import { MapLibreAssertable, MaplibreCypressHelper } from "./maplibre-helper";
 
-export class MapLibreAssertable<T> extends Assertable<T> {
-  private comparePixels = (
-    buffer1: Buffer,
-    buffer2: Buffer,
-    width: number,
-    height: number
-  ) => pixelmatch(buffer1, buffer2, null, width, height);
-
-  private compareSnapshots = (
-    snapshot: Cypress.Chainable<{
-      buffer: Buffer;
-      height: number;
-      width: number;
-    }>
-  ) => {
-    const snapshotChainable = this.chainable as unknown as Cypress.Chainable<{
-      buffer: Buffer;
-      height: number;
-      width: number;
-    }>;
-    return then(
-      snapshotChainable.then(
-        (subject: { buffer: Buffer; height: number; width: number }) =>
-          snapshot.then(({ buffer, width, height }) =>
-            cy.wrap(this.comparePixels(buffer, subject.buffer, width, height))
-          )
-      )
-    );
-  };
-
-  public shouldEqualSnapshot = (
-    snapshot: Cypress.Chainable<{
-      buffer: Buffer;
-      height: number;
-      width: number;
-    }>
-  ) => this.compareSnapshots(snapshot).shouldEqual(0);
-
-  public shouldNotEqualSnapshot = (
-    snapshot: Cypress.Chainable<{
-      buffer: Buffer;
-      height: number;
-      width: number;
-    }>
-  ) => this.compareSnapshots(snapshot).shouldBeGreaterThen(0);
-}
 export class E2eDriver {
   private helper = new MaplibreCypressHelper();
 
@@ -59,7 +10,7 @@ export class E2eDriver {
   };
 
   given = {
-    alertStub: () => cy.on("window:alert", this.helper.given.stub("alert")),
+    alertStub: () => this.helper.given.onAlert(this.helper.given.stub("alert")),
   };
   when = {
     wait: (ms: number) => this.helper.when.wait(ms),
